Fix garbage transcripts: use fp32 (not fp16) Whisper on WebGPU
fp16 Whisper decoding is numerically unstable on many GPUs and degenerated
into repeated punctuation instead of text. Switch WebGPU dtype to fp32 (the
reference precision) — verified end-to-end: a 5s clip now transcribes
correctly at full accuracy. Also default to the lighter tiny.en checkpoint
for a faster, smaller first run.

+++ src/worker.ts
@@ -71,7 +71,10 @@ async function getTranscriber(model: ModelId): Promise<AutomaticSpeechRecognitio
   ) => Promise<AutomaticSpeechRecognitionPipeline>;
   transcriber = await build('automatic-speech-recognition', model, {
     device,
-    dtype: device === 'webgpu' ? 'fp16' : 'q8',
+    // fp32 on WebGPU: fp16 Whisper decoding is numerically unstable on many
+    // GPUs and degenerates into repeated punctuation ("​ ​ ​…"). fp32 is the
+    // reference precision and transcribes correctly; q8 keeps the CPU path light.
+    dtype: device === 'webgpu' ? 'fp32' : 'q8',
     progress_callback: (p: {
       status: string;
       file?: string;
